test_attr: user-friendly type list
Checking the allowed types for attributes such as `timeOffset` gave
a suboptimal user-hint of the form:
```
Error: Attribute timeOffset in /data/0/particles/electron/positionOffset`
is not of type '<map object at 0x11aa73050>' (is 'float32')!`
```

Let's concat the types properly to
```
Error: Attribute timeOffset in /data/0/particles/electron/positionOffset`
is not of type 'float32' or 'float64' (is 'str')!
```

Thanks for the report, @rfbird! :sparkles: